requirements.txt
google-genai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
pandas
plotly
kaleido
streamlit
fpdf2
schedule